Add __builtin_stack_top
When __builtin_frame_address is used to retrieve the address of the
function stack frame, the frame pointer register is required, which
wastes one register and 2 instructions.  For x86-32, one less register
means significant negative impact on performance.  This patch adds a
new builtin function, __builtin_stack_top, to x86 backend.  It returns
the stack address when the function is called.

gcc/

	PR target/66960
	* config/i386/i386.c (ix86_expand_prologue): Sorry if DRAP is
	used and the stack address has been taken.
	(ix86_builtins): Add IX86_BUILTIN_STACK_TOP.
	(ix86_init_mmx_sse_builtins): Add __builtin_stack_top.
	(ix86_expand_builtin): Handle IX86_BUILTIN_STACK_TOP.
	* config/i386/i386.h (machine_function): Add stack_top_taken.
	* doc/extend.texi: Document __builtin_stack_top.

gcc/testsuite/

	PR target/66960
	* gcc.target/i386/pr66960-1.c: New test.
	* gcc.target/i386/pr66960-2.c: Likewise.
	* gcc.target/i386/pr66960-3.c: Likewise.
	* gcc.target/i386/pr66960-4.c: Likewise.
	* gcc.target/i386/pr66960-5.c: Likewise.