Generate indirect branch relocation via GOT
This patch extends -fno-plt to non-PIC on x86.  -fno-plt works in 64-bit
mode with the existing binutils.  For 32-bit, we need the updated
assembler and linker to support "call/jmp *foo@GOT" with a new relocation
different from R_386_GOT32 to indicate that this relocation applies to
indirect branches via GOT.

gcc/

	* config/i386/i386.c (ix86_nopic_noplt_attribute_p): Check
	HAVE_LD_IX86_INDBR_RELOC == 0 before returning false.
	(ix86_output_call_insn): Generate "%!jmp/call\t*%p0@GOT" for
	32-bit.

gcc/testsuite/

	* gcc.target/i386/pr66232-10.c: New file.
	* gcc.target/i386/pr66232-11.c: Likewise.
	* gcc.target/i386/pr66232-12.c: Likewise.
	* gcc.target/i386/pr66232-13.c: Likewise.
	* lib/target-supports.exp (check_effective_target_indbr_reloc):
	New.