1) Pass "Qualified storage" everywhere.
Storage type is the type that is actually storing the alternatives; it's
never cv-qualified or a reference.

Qualified storage is a cv-qualified reference of a storage, indicating
extra qualification for the values extracted from the storage.

Usually the qualification comes from the variant.

For example, for `cnost variant<int&&>&`:
Alternative type is int&&;
Storage type is _Reference_storage<int&&>;
Qualified reference is const _Reference_storage<int&&>&.

To extract the "user desired type", simple recover the alternative type
from the storage type (they are one-to-one mappings), then apply the
qualification, in this case `const (int&&) &`, aka `const int&`.

2) FUN(T_i) behavior for variant(T&&) and operator=(T&&).